Move `minimist.js`
